Add GraphicsBufferV2 Part 1
This will replace the old per object buffers which are inefficient and causes issues with raytracing renderer.
* GraphicsBuffer will be deprecated.
* fix unittests, glfw window now will not pop open
* Replace vertex and index buffer with GraphiscBufferV2
* Add UT for GraphicsBufferV2
* Add buffer mapper (maps object id to each buffer's offset for each object)

+++ src/graphics_engine/raytracing.ipp
@@ -582,7 +582,7 @@ void GraphicsEngineRayTracing<GraphicsEngineT>::create_shader_binding_table()
 
 	const auto align_up = [](uint32_t val, uint32_t alignment) -> uint32_t
 	{
-		return uint32_t((val + (uint32_t(alignment) - 1)) & ~uint32_t(alignment - 1));
+		return (val + alignment - 1) & ~(alignment - 1);
 	};
 
 	// sets the following strides and sizes for each group
